requirements.txt
streamlit==1.35.0
requests==2.31.0
pymongo==4.7.2
bcrypt
python-dotenv
pandas
tabulate
dnspython==2.6.1
beautifulsoup4==4.12.3
lxml

